% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyte_names_fix.R
\name{analyte_names_fix}
\alias{analyte_names_fix}
\title{Fix analyte names}
\usage{
analyte_names_fix(analyte_reference = analyte_reference, dta = dta)
}
\arguments{
\item{analyte_reference}{From external table or database}

\item{dta}{returned data from clean colnames function}
}
\description{
Fix analyte names
}
