% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyte_names_fix.R
\name{analyte_names_fix2}
\alias{analyte_names_fix2}
\title{Fix analyte names}
\usage{
analyte_names_fix2(analyte_reference_path = analyte_reference_path, dta = dta)
}
\arguments{
\item{analyte_reference_path}{reference path}

\item{dta}{dta path from colname_clean function}
}
\description{
Fix analyte names
}
