% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{cor_plot}
\alias{cor_plot}
\title{Visualise correlogram - analytes}
\usage{
cor_plot(
  dta,
  analyte_col = "analyte",
  conc_col = "obs_conc",
  cor_type = "spearman"
)
}
\arguments{
\item{dta}{dataframe from pipeline}

\item{analyte_col}{column name for analyte in "quotes"}

\item{conc_col}{column name for observed concentration in "quotes"}

\item{cor_type}{character, type of correlation to compute}
}
\value{
A graph (correlogram) of correlations
}
\description{
Visualise correlogram - analytes
}
