% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{correlation}
\alias{correlation}
\title{Correlation tests}
\usage{
correlation(dta, cor_type = "spearman")
}
\arguments{
\item{dta}{dataframe from the pipeline}

\item{cor_type}{type of correlation in string format. eg "pearson"}
}
\value{
a flatten correlation matrix (table)
}
\description{
Correlation tests
}
