% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{data_import}
\alias{data_import}
\title{data_import function Reads in valid files, one line per well, keeping standards,
controls and unknowns in one data frame called "dta"}
\usage{
data_import(dir_data)
}
\arguments{
\item{dir_data}{directory of datasets}
}
\value{
None
}
\description{
data_import function Reads in valid files, one line per well, keeping standards,
controls and unknowns in one data frame called "dta"
}
