% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_split.R
\name{data_split}
\alias{data_split}
\title{Split into participant and QC data}
\usage{
data_split(dta = dta)
}
\arguments{
\item{dta}{data from symbols remove function.}
}
\description{
Split into participant and QC data
}
