% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename_separate.R
\name{dates_filename}
\alias{dates_filename}
\title{Check date format in FileName}
\usage{
dates_filename(d)
}
\arguments{
\item{d}{Dataset created by preceding function}
}
\description{
Check date format in FileName
}
