% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename_separate.R
\name{filename_separate}
\alias{filename_separate}
\title{Extract relevant metadata from filenames of valid imported Luminex files}
\usage{
filename_separate(data = dta, instrument_names = instrument_names)
}
\arguments{
\item{data}{Output of the \code{data_import()} function.
A valid Luminex tibble or dataframe.}

\item{instrument_names}{Character vector of expected instrument names.
This is sourced from the accompanying metadata from the experiment}
}
\value{
A tibble of the input data with added columns for extracted metadata
e.g., date, kit, instrument, plate and rerun.
}
\description{
Extract useful metadata e.g., (kit name, instrument name, date) from
the \code{filenames} column created by the \code{data_import()} function
and assign each metadata type a column in the output tibble.
}
\details{
As part of the standard operating procedures (SOP), The raw Luminex
\code{.txt} files are named with metadata with a specified guideline. This
function extracts these metadata (from the \code{filenames} column created
by the \code{data_import()} function) and adds new columns to
the output tibble for each type of metadata e.g., "date", "instrument_name" etc.
i.e, values of these new columns are extracted from \code{filename} column of the
input data.

The new columns are parsed with its correct data types. e.g., dates are parsed
with date-time format.

The \code{filename_separate()} function works with the assumption that its
input data is available in the R environment. If input data is not already
available in the environment, ensure that it is loaded in the environment
}
\examples{
#assign instrument names to a variable called ins_names
ins_names <- c("bp", "rd")

#read input data to the environment
my_data <- readRDS("tests/testthat/rds/dta_import.rds")

filename_separate(data = my_data, instrument_names = ins_names)

}
