% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename_separate.R
\name{filename_separate1}
\alias{filename_separate1}
\title{filename_separate1 function not exported.}
\usage{
filename_separate1(data = d)
}
\arguments{
\item{data}{dta file from data_import function. it is assumed that the input data is already loaded in the environment.
if input data is not already in the environment, ensure that it is loaded or read it into the environment.}
}
\description{
filename_separate1 function not exported.
}
