% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{grp_test}
\alias{grp_test}
\title{Comparison between groups being analysed}
\usage{
grp_test(dta_prtc, dta_clin)
}
\arguments{
\item{dta_prtc}{participant data}

\item{dta_clin}{clinical data}
}
\value{
A dataframe of comparison between groups summary
}
\description{
Comparison between groups being analysed
}
