% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{histogram_plot}
\alias{histogram_plot}
\title{Histogram plot}
\usage{
histogram_plot(
  dta_prtc,
  analyte_col = "analyte",
  conc_col = "obs_conc",
  facet_rows = 5
)
}
\arguments{
\item{dta_prtc}{processed dataframe}

\item{analyte_col}{column name for analyte in "quotes"}

\item{conc_col}{column name for observed concentration in "quotes"}

\item{facet_rows}{numeric value indicating number of rows to facet.}
}
\value{
a plot of histogram
}
\description{
Histogram plot
}
