% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename_separate.R
\name{instrument_name_check}
\alias{instrument_name_check}
\title{Check whether instrument_name is same as input instrument_name.}
\usage{
instrument_name_check(d, instrument_names)
}
\arguments{
\item{d}{dataframe created by preceding function}

\item{instrument_names}{character vector from essential metadata}
}
\description{
Check whether instrument_name is same as input instrument_name.
}
