% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{norm_tab}
\alias{norm_tab}
\title{Test for normality}
\usage{
norm_tab(dat)
}
\arguments{
\item{dat}{data to generate test of normality table from}
}
\value{
a table of normality test
}
\description{
Test for normality
}
