% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_config.R
\name{pipeline_config}
\alias{pipeline_config}
\title{pipeline_config}
\usage{
pipeline_config(
  wd,
  dd,
  md = NA,
  aref,
  isa_path = NULL,
  config_isa = FALSE,
  tech_reps
)
}
\arguments{
\item{wd}{work directory}

\item{dd}{data directory}

\item{md}{NA}

\item{aref}{analyte reference}

\item{isa_path}{path to isa}

\item{config_isa}{NA}

\item{tech_reps}{technical replicates}
}
\description{
pipeline_config
}
