% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drL_plotLum.R
\name{plotQQnorm}
\alias{plotQQnorm}
\title{Plot qqnorm of the residuals for each analyte}
\usage{
plotQQnorm(x, subset.plot = NULL, psize = 1.8, ncol = NULL, nrow = NULL, ...)
}
\arguments{
\item{x}{a \code{scluminex} object}

\item{subset.plot}{list of analytes to be plotted. By default
plot all analytes.}

\item{psize}{point size}

\item{ncol}{number of columns to plot the analytes.}

\item{nrow}{number of rows to plot the analytes.}

\item{...}{other \code{ggplot} arguments}
}
\value{
A \code{ggplot} object
}
\description{
Plot qqnorm of the residuals for each analyte the  \code{scluminex}
object in a ggplot format so new parameters can be added
}
