% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{render_report}
\alias{render_report}
\title{render statistical summary report}
\usage{
render_report(dta)
}
\arguments{
\item{dta}{rds file with its extension}
}
\value{
an html report
}
\description{
render statistical summary report
}
