% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_participant_report.R
\name{report_isa}
\alias{report_isa}
\title{Create metadata report and compare to ISA data}
\usage{
report_isa(dta_list = dta_list, isa_path = isa_path, config_isa = config_isa)
}
\arguments{
\item{dta_list}{data list from data_split function}

\item{isa_path}{path to ISA files}

\item{config_isa}{path to config ISA}
}
\value{
isa_report_list.rds
}
\description{
Create metadata report and compare to ISA data
}
