% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_raw.R
\name{save_raw}
\alias{save_raw}
\title{Save raw data file}
\usage{
save_raw(dta = dta)
}
\arguments{
\item{dta}{dataset from analyte_name_fix2 function}
}
\value{
a file named dta_raw.rds
}
\description{
Save raw data file
}
