% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{summ_table}
\alias{summ_table}
\title{Summary table}
\usage{
summ_table(dta, analyte_col = "analyte", conc_col = "obs_conc")
}
\arguments{
\item{dta}{processed dataframe}

\item{analyte_col}{column name for analyte in "quotes"}

\item{conc_col}{column name for observed concentrations in "quotes"}
}
\value{
a table of summaries
}
\description{
Summary table
}
