% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol_remove.R
\name{symbols_remove}
\alias{symbols_remove}
\title{Remove symbols from concentration and FI columns}
\usage{
symbols_remove(dta = dta)
}
\arguments{
\item{dta}{data returned from previous function - analyte_names_fix}
}
\description{
Remove symbols from concentration and FI columns
}
