% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames_clean.R
\name{colnames_clean}
\alias{colnames_clean}
\title{Clean column names}
\usage{
colnames_clean(dta)
}
\arguments{
\item{dta}{output data of the preceding function ie., filename_separate function. The colnames_clean function assumes its input data is already loaded in the environment.
if input data is not already available in the environment, ensure that it is loaded or read it into the environment.}
}
\value{
a tibble of the input data but with cleaned column names eg. lower case column names, removal of special characters in column names
}
\description{
Clean column names
}
