% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{correlation}
\alias{correlation}
\title{Correlation tests}
\usage{
correlation(
  dta,
  analyte_col = "analyte",
  conc_col = "obs_conc",
  cor_type = "spearman"
)
}
\arguments{
\item{dta}{dataframe from the pipeline}

\item{analyte_col}{column name for analyte in "quotes"}

\item{conc_col}{column name for observed concentration in "quotes"}

\item{cor_type}{type of correlation in string format. eg "pearson"}
}
\value{
a flatten correlation matrix (table)
}
\description{
Correlation tests
}
