% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename_separate.R
\name{filename_separate}
\alias{filename_separate}
\title{Filename separation
Check structure of observations in the filename column in the imported data}
\usage{
filename_separate(data = dta, instrument_names = instrument_names)
}
\arguments{
\item{data}{Output from the data_import function. The filename_separate function works with the assumption that this input data is already loaded in the environment.
if input data is not already in the environment, ensure that it is loaded or read it into the environment.}

\item{instrument_names}{Character vector of expected instrument names. This is sourced from the accompanying metadata from the experiment}
}
\value{
a tibble of the input data with added columns for date, kit, instrument, plate and rerun. values of these new columns are extracted from "filename" column of the input data
}
\description{
Filename separation
Check structure of observations in the filename column in the imported data
}
