% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{histogram_plot}
\alias{histogram_plot}
\title{Histogram plot}
\usage{
histogram_plot(dta_prtc, facet_rows = 5)
}
\arguments{
\item{dta_prtc}{processed dataframe}

\item{facet_rows}{numeric value indicating number of rows to facet.}
}
\value{
a plot of histogram
}
\description{
Histogram plot
}
