% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{qq_plot}
\alias{qq_plot}
\title{Generate quantile-quantile plots to assess normality}
\usage{
qq_plot(dta, facet_row = 5)
}
\arguments{
\item{dta}{dataframe from pipeline}

\item{facet_row}{numeric value indicating number of rows to facet}
}
\value{
A quantile-quantil graph
}
\description{
Generate quantile-quantile plots to assess normality
}
